Python/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-ai-textanalytics
openai
openai[embeddings]
azure-core
numpy
datetime
matplotlib
azure-storage-blob>=12.0.0
pandas
azure-search-documents

# English
#https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.1.0/en_core_web_lg-3.1.0.tar.gz#egg=en_core_web_lg

# German 
#https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-3.1.0/de_core_news_lg-3.1.0.tar.gz#egg=de_core_news_lg
